// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/adt/shared/DTOs/adt-bed-reservation-admitted-patient-information.dto.ts
export class ADT_BedReservationAdmittedPatientInformation_DTO {
  PatientId: number = 0;
  PatientVisitId: number = 0;
  PatientName: string = "";
  VisitCode: string = "";
  AdmittedDate: string = "";
  AgeSex: string = "";
  PatientCode: string = "";
  CurrentWardBed: string = "";
  BedInformation: unknown;
  DateOfBirth: string = "";
  Gender: string = "";
}